+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.5 (2020-12-18)
+### Optimisations
+- Prevent double mutations from DOM listeners
+- more comments on source code
+- better scope handling in dom.js
+- function for core setter
 ## 0.0.4 (2020-12-18)
 ### Features
 - demo page, input connect